+++ src/main/java/com/epam/reportportal/testng/TestNGService.java
@@ -59,6 +59,7 @@ import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 import static com.epam.reportportal.testng.util.ItemTreeUtils.createKey;
 import static java.util.Optional.ofNullable;
@@ -73,7 +74,11 @@ import static org.apache.commons.lang3.exception.ExceptionUtils.getStackTrace;
 public class TestNGService implements ITestNGService {
 
 	private static final String AGENT_PROPERTIES_FILE = "agent.properties";
-	private static final Predicate<StackTraceElement> IS_RETRY_ELEMENT = e -> "org.testng.internal.TestInvoker".equals(e.getClassName())
+	private static final Set<String> TESTNG_INVOKERS = Stream.of(
+			"org.testng.internal.TestInvoker",
+			"org.testng.internal.invokers.TestInvoker"
+	).collect(Collectors.toSet());
+	private static final Predicate<StackTraceElement> IS_RETRY_ELEMENT = e -> TESTNG_INVOKERS.contains(e.getClassName())
 			&& "retryFailed".equals(e.getMethodName());
 	private static final Predicate<StackTraceElement[]> IS_RETRY = eList -> Arrays.stream(eList).anyMatch(IS_RETRY_ELEMENT);
 	private static final int MAXIMUM_HISTORY_SIZE = 1000;
@@ -132,8 +137,8 @@ public class TestNGService implements ITestNGService {
 	/**
 	 * Return current instance of {@link ReportPortal} class
 	 *
-	 * @deprecated use <code>Launch.currentLaunch().getClient()</code>
 	 * @return ReportPortal instance
+	 * @deprecated use <code>Launch.currentLaunch().getClient()</code>
 	 */
 	@Deprecated
 	public static ReportPortal getReportPortal() {
@@ -143,8 +148,8 @@ public class TestNGService implements ITestNGService {
 	/**
 	 * Set current instance of {@link ReportPortal} class
 	 *
-	 * @deprecated use {@link TestNGService#TestNGService(com.epam.reportportal.service.ReportPortal)}
 	 * @param reportPortal class instance
+	 * @deprecated use {@link TestNGService#TestNGService(com.epam.reportportal.service.ReportPortal)}
 	 */
 	@Deprecated
 	protected static void setReportPortal(ReportPortal reportPortal) {
@@ -550,18 +555,15 @@ public class TestNGService implements ITestNGService {
 	@Nonnull
 	protected StartTestItemRQ buildStartTestItemRq(@Nonnull ITestContext testContext) {
 		StartTestItemRQ rq = new StartTestItemRQ();
-		XmlTest currentXmlTest = testContext.getCurrentXmlTest();
-		if (currentXmlTest != null) {
-			List<XmlClass> xmlClasses = currentXmlTest.getXmlClasses();
-			if (xmlClasses != null) {
-				XmlClass xmlClass = xmlClasses.get(0);
-				if (xmlClass != null) {
-					rq.setCodeRef(xmlClass.getName());
-					ofNullable(xmlClass.getSupportClass()).map(c -> c.getAnnotation(Attributes.class))
-							.ifPresent(a -> rq.setAttributes(AttributeParser.retrieveAttributes(a)));
-				}
-			}
-		}
+		Set<ItemAttributesRQ> attributes = rq.getAttributes() == null ? new HashSet<>() : new HashSet<>(rq.getAttributes());
+		rq.setAttributes(attributes);
+		ofNullable(testContext.getCurrentXmlTest()).map(XmlTest::getXmlClasses).ifPresent(xmlClasses -> xmlClasses.forEach(xmlClass -> {
+			String className = xmlClass.getName();
+			String codeRef = rq.getCodeRef();
+			rq.setCodeRef(codeRef == null ? className : codeRef + ";" + className);
+			ofNullable(xmlClass.getSupportClass()).map(c -> c.getAnnotation(Attributes.class))
+					.ifPresent(a -> attributes.addAll(AttributeParser.retrieveAttributes(a)));
+		}));
 		rq.setName(testContext.getName());
 		rq.setStartTime(testContext.getStartDate());
 		rq.setType("TEST");
@@ -648,22 +650,21 @@ public class TestNGService implements ITestNGService {
 	 * @param testResult TestNG's testResult context
 	 * @return Step Parameters being sent to Report Portal
 	 */
-	private List<ParameterResource> createAnnotationParameters(ITestResult testResult) {
-		Parameters parametersAnnotation = getMethodAnnotation(Parameters.class, testResult);
-		if (parametersAnnotation == null) {
-			return Collections.emptyList();
-		}
-		String[] keys = parametersAnnotation.value();
-		Object[] parameters = testResult.getParameters();
-		if (parameters.length != keys.length || keys.length <= 0) {
-			return Collections.emptyList();
-		}
-		return IntStream.range(0, keys.length).mapToObj(i -> {
-			ParameterResource parameter = new ParameterResource();
-			parameter.setKey(keys[i]);
-			parameter.setValue(parameters[i] == null ? NULL_VALUE : parameters[i].toString());
-			return parameter;
-		}).collect(toList());
+	@Nonnull
+	private List<ParameterResource> createAnnotationParameters(@Nonnull ITestResult testResult) {
+		return getMethodAnnotation(Parameters.class, testResult).map(a -> {
+			String[] keys = a.value();
+			Object[] parameters = testResult.getParameters();
+			if (parameters.length != keys.length || keys.length <= 0) {
+				return Collections.<ParameterResource>emptyList();
+			}
+			return IntStream.range(0, keys.length).mapToObj(i -> {
+				ParameterResource parameter = new ParameterResource();
+				parameter.setKey(keys[i]);
+				parameter.setValue(parameters[i] == null ? NULL_VALUE : parameters[i].toString());
+				return parameter;
+			}).collect(toList());
+		}).orElse(Collections.emptyList());
 	}
 
 	/**
@@ -674,15 +675,16 @@ public class TestNGService implements ITestNGService {
 	 * @param testResult TestNG's testResult context
 	 * @return Step Parameters being sent to ReportPortal
 	 */
-	private List<ParameterResource> createDataProviderParameters(ITestResult testResult) {
-		Test testAnnotation = getMethodAnnotation(Test.class, testResult);
-		Method method = getMethod(testResult);
-		Object[] parameters = testResult.getParameters();
-		if (method == null || testAnnotation == null || isBlank(testAnnotation.dataProvider()) || parameters == null
-				|| parameters.length <= 0) {
-			return Collections.emptyList();
-		}
-		return ParameterUtils.getParameters(method, Arrays.asList(parameters));
+	@Nonnull
+	private List<ParameterResource> createDataProviderParameters(@Nonnull ITestResult testResult) {
+		return getMethodAnnotation(Test.class, testResult).map(a -> {
+			Method method = getMethod(testResult);
+			Object[] parameters = testResult.getParameters();
+			if (method == null || isBlank(a.dataProvider()) || parameters == null || parameters.length <= 0) {
+				return Collections.<ParameterResource>emptyList();
+			}
+			return ParameterUtils.getParameters(method, Arrays.asList(parameters));
+		}).orElse(Collections.emptyList());
 	}
 
 	private List<ParameterResource> crateFactoryParameters(ITestResult testResult) {
@@ -797,29 +799,26 @@ public class TestNGService implements ITestNGService {
 	 * @param testResult Where to find
 	 * @return test item ID or null
 	 */
-	private String extractUniqueID(ITestResult testResult) {
-		UniqueID itemUniqueID = getMethodAnnotation(UniqueID.class, testResult);
-		return itemUniqueID != null ? itemUniqueID.value() : null;
+	@Nullable
+	private String extractUniqueID(@Nonnull ITestResult testResult) {
+		return getMethodAnnotation(UniqueID.class, testResult).map(UniqueID::value).orElse(null);
 	}
 
+	@Nullable
 	private TestCaseIdEntry getTestCaseId(@Nonnull String codeRef, @Nonnull ITestResult testResult) {
-		TestCaseId testCaseId = getMethodAnnotation(TestCaseId.class, testResult);
 		Method method = getMethod(testResult);
 		List<Object> parameters = ofNullable(testResult.getParameters()).map(Arrays::asList).orElse(null);
-		TestCaseIdEntry id = ofNullable(method).map(m -> TestCaseIdUtils.getTestCaseId(testCaseId, m, parameters))
+		TestCaseIdEntry id = getMethodAnnotation(TestCaseId.class,
+				testResult
+		).flatMap(a -> ofNullable(method).map(m -> TestCaseIdUtils.getTestCaseId(a, m, parameters)))
 				.orElse(TestCaseIdUtils.getTestCaseId(codeRef, parameters));
-		if (id == null) {
-			return null;
-		}
-		return id.getId().endsWith("[]") ? new TestCaseIdEntry(id.getId().substring(0, id.getId().length() - 2)) : id;
+
+		return id == null ? null : id.getId().endsWith("[]") ? new TestCaseIdEntry(id.getId().substring(0, id.getId().length() - 2)) : id;
 	}
 
-	protected Set<ItemAttributesRQ> createStepAttributes(ITestResult testResult) {
-		Attributes attributesAnnotation = getMethodAnnotation(Attributes.class, testResult);
-		if (attributesAnnotation != null) {
-			return AttributeParser.retrieveAttributes(attributesAnnotation);
-		}
-		return null;
+	@Nullable
+	protected Set<ItemAttributesRQ> createStepAttributes(@Nonnull ITestResult testResult) {
+		return getMethodAnnotation(Attributes.class, testResult).map(AttributeParser::retrieveAttributes).orElse(null);
 	}
 
 	@Nullable
@@ -838,12 +837,9 @@ public class TestNGService implements ITestNGService {
 	 * @param testResult Where to find
 	 * @return {@link Annotation} or null if doesn't exists
 	 */
-	private <T extends Annotation> T getMethodAnnotation(Class<T> annotation, ITestResult testResult) {
-		Method method = getMethod(testResult);
-		if (null != method) {
-			return method.getAnnotation(annotation);
-		}
-		return null;
+	@Nonnull
+	private <T extends Annotation> Optional<T> getMethodAnnotation(@Nonnull Class<T> annotation, @Nonnull ITestResult testResult) {
+		return ofNullable(getMethod(testResult)).map(m -> m.getAnnotation(annotation));
 	}
 
 	/**